requirements.txt
asgiref==3.9.1
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.10.0
cffi==1.15.1
constantly==15.1.0
cryptography==43.0.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==5.2.6
django-bootstrap3==21.1
django-bootstrap4==21.1
django-cors-headers==3.13.0
django-unfold==0.65.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==2.1.2
incremental==22.10.0
Pillow==9.3.0
psycopg2==2.9.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==24.2.1
python-dotenv==0.21.1
pytz==2021.3
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
static3==0.7.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.0.1
tzdata==2025.2
whitenoise==6.5.0
zipp==3.6.0
zope.interface==6.4.post2
